requirements.txt
pyserial
typer[all]
pytest
pytest-mock